Add keyword segment taxonomy, tweak robustness of taxonomy importing
