+++ server/src/server.ts
@@ -5,7 +5,7 @@ import * as express from 'express';
 import * as Sentry from '@sentry/node';
 import { NextFunction, Request, Response } from 'express';
 import { importLocales } from './lib/model/db/import-locales';
-import { importTaxonomies } from './lib/model/db/import-taxonomies';
+import { importTargetSegments } from './lib/model/db/import-target-segments';
 import Model from './lib/model';
 import {
   getFullClipLeaderboard,
@@ -288,7 +288,7 @@ export default class Server {
       if (doImport) {
         await importSentences(await this.model.db.mysql.createPool());
       }
-      await importTaxonomies();
+      await importTargetSegments();
       this.print('Maintenance complete');
     } catch (err) {
       this.print('Maintenance error', err);
